split message and messages source, add kustorestrecord
